% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.vblogitfit}
\alias{logLik.vblogitfit}
\title{Log-evidence}
\usage{
\method{logLik}{vblogitfit}(x, ...)
}
\arguments{
\item{x}{object from vblogit}

\item{...}{ignored.}
}
\description{
Log-evidence
}
