% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.vblogitfit}
\alias{plot.vblogitfit}
\title{vblogit fit plot}
\usage{
\method{plot}{vblogitfit}(x, ...)
}
\arguments{
\item{x}{object from vblogit}

\item{...}{passed to plot.vblogitfit_marginals.}
}
\description{
Plot something.
}
