% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.vblogitfit_marginals}
\alias{plot.vblogitfit_marginals}
\title{Marginals plot}
\usage{
\method{plot}{vblogitfit_marginals}(x, ...)
}
\arguments{
\item{x}{object from vblogit}

\item{...}{ignored.}
}
\description{
Marginals plot
}

