% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.vblogitfit}
\alias{print.vblogitfit}
\title{vblogit fit print method}
\usage{
\method{print}{vblogitfit}(x, ...)
}
\arguments{
\item{x}{vblogit object}

\item{...}{ignored.}
}
\description{
vblogit fit print method
}
