% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.vblogitfit_marginals}
\alias{print.vblogitfit_marginals}
\title{Marginals print}
\usage{
print.vblogitfit_marginals(x, ...)
}
\arguments{
\item{x}{object from vblogit}

\item{...}{ignored.}
}
\description{
Marginals print
}

