% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.vblogitfit}
\alias{summary.vblogitfit}
\title{vblogit fit summary method
Summary method for logistic regression with approximately Normal posterior}
\usage{
\method{summary}{vblogitfit}(object, ...)
}
\arguments{
\item{...}{ignored.}

\item{x}{vblogit object}
}
\description{
vblogit fit summary method
Summary method for logistic regression with approximately Normal posterior
}
