\name{vblogistic-package}
\alias{vblogistic-package}
\alias{vblogistic}
\docType{package}
\title{
Bayesian logistic regression via variational approximation
}
\description{
Fit the logistic regression model using the tangential method described by Jaakkola and Jordan 1996. 
}
\details{
The main function is \link{vblogit}. 
}
\author{
Maintainer: Tuomas Rajala <tuomas.rajala@iki.fi>
}
\references{
Jaakkola and Jordan (1996): A variational approach to Bayesian logistic regression models and their extensions
}
\keyword{ package }
\seealso{
Default R-implementation is given by \link{glm}
}

