% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formula.R
\name{vblogit.fmla}
\alias{vblogit.fmla}
\title{Formula interface for vblogit used by the logi.engine}
\usage{
vblogit.fmla(formula, data, offset, subset, method = "vblogit.fit", ...)
}
\description{
Formula interface for vblogit used by the logi.engine
}

