% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_dense.R
\name{vblogit_dense}
\alias{vblogit_dense}
\title{vb-logit basic
No sparse matrix class}
\usage{
vblogit_dense(
  y,
  X,
  offset,
  eps = 0.01,
  m0,
  S0,
  S0i,
  xi0,
  verb = FALSE,
  maxiter = 1000,
  ...
)
}
\description{
vb-logit basic
No sparse matrix class
}
