% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logistic_regression_slow_2.R
\name{vblogit_slow2}
\alias{vblogit_slow2}
\title{vb-logit basic
No sparse matrix class}
\usage{
vblogit_slow2(y, X, offset, eps = 0.01, m0, S0, S0i, xi0, verb = FALSE,
  maxiter = 1000, ...)
}
\description{
vb-logit basic
No sparse matrix class
}

