% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.vblogitfit}
\alias{vcov.vblogitfit}
\title{Var-Cov Method}
\usage{
\method{vcov}{vblogitfit}(object, ...)
}
\arguments{
\item{object}{'vblogitfit' object}

\item{...}{not used}
}
\description{
Var-Cov Method
}
