\name{marginals.vblogifit}
\alias{marginals.vblogifit}
\title{Posterior marginals of the parameters in a vblogit fit}
\usage{
marginals.vblogifit(x, log = TRUE, ...)
}
\arguments{
  \item{x}{vblogit object}

  \item{log}{If FALSE, exponentiate the values.}

  \item{...}{ignored}
}
\description{
Use the fact that the variables are gaussian.
}

