% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_marginals.R
\name{marginals.vblogifit}
\alias{marginals.vblogifit}
\title{Posterior marginals of the parameters in a vblogit fit}
\usage{
marginals.vblogifit(x, which = NULL, log = TRUE, ...)
}
\arguments{
\item{x}{vblogit object}

\item{which}{If NULL, create all marginals. Otherwise the indices to create.}

\item{log}{If FALSE, exponentiate the values.}

\item{...}{ignored}
}
\description{
Use the fact that the variables are gaussian.
}
