% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.vblogitfit_marginals}
\alias{plot.vblogitfit_marginals}
\title{Marginals plot}
\usage{
\method{plot}{vblogitfit_marginals}(x, ncol = 2, ...)
}
\arguments{
\item{x}{object from vblogit}

\item{ncol}{default:2}

\item{...}{ignored.}
}
\description{
Marginals plot
}
