% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sum_vblogitfit}
\alias{print.sum_vblogitfit}
\title{Print for vblogit fit summary}
\usage{
\method{print}{sum_vblogitfit}(x, ...)
}
\arguments{
\item{x}{vblogit object}

\item{...}{ignored.}
}
\description{
Print for vblogit fit summary
}
