% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.vblogitfit}
\alias{print.vblogitfit}
\title{vblogit fit print method}
\usage{
print.vblogitfit(x, ...)
}
\arguments{
\item{x}{vblogit object}

\item{...}{ignored.}
}
\description{
vblogit fit print method
}

